Disallow recursive definitions or free variables that have no corresponding top-level definition
Before adding a definition, collect the expression's set of free variables. Fail if any free
variable cannot be found in the current set of top-level definitions.

Given the default definitions:

  i = λx. x
  k = λx. λy. x
  fix = λf. (λx. f (x x)) (λy. f (y y))

The following definition is disallowed because it refers to g, which is has no top-level definition:

  f = \x. g x

The following definition is disallowed because it is recursive:

  m = \x. m x

The following definition is disallowed because it appears to be recursive, but would really replace
an existing definition if accepted:

  i = i x

When a user deletes a definition, we should warn if any remaining definitions still reference it.