Fix treatment of bound variables when evaluating under a lambda
Whenever we go under a lambda, we make the bound var free. We also
delete that name from the global environment. This is safe since if we
were evaluating the lambda normally, it would shadow any global
variable. Once we've evaluated the body of the lambda, we re-bind the
free variable, which may have moved, and thus have a different index
than before.