Add Stash representing suspended computations
A Stash is a stack of Stacks. We currently only use this to evaluate the
operand of a stuck application, but it can also be used to scrutinize
structured data (which I may or may not add). This is also called the
Dump in the literature. I just like that Stash looks like Stack (and we
are stashing Stacks away for later).