Split Globals out into its own module
Also move instantiate into Node and make Node depend on Globals.
Globals.add now takes a procedure from unit to the Heap node type so
that we can tie a recursive know in the case of self-recursion at the
global level. Globals.add reserves a pointer for the new global and
associates the name to it before node is actually created. Sort of
overkill, but it lets us keep Machine.purs from even knowing about
the Expr constructors.